@@ -3,6 +3,8 @@
 namespace App\Entity;
 
 use App\Repository\PostsListRepository;
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 
 #[ORM\Entity(repositoryClass: PostsListRepository::class)]
@@ -13,35 +15,26 @@ class PostsList
     #[ORM\Column(type: 'integer')]
     private $id;
 
-    #[ORM\Column(type: 'string', length: 255)]
-    private $post_name;
+    #[ORM\Column]
+    private array $post_name = [];
 
-    #[ORM\Column(type: 'string', length: 255)]
-    private $post_url;
+    #[ORM\Column]
+    private array $post_content = [];
 
-    #[ORM\Column(type: 'string', length: 255)]
-    private $post_id;
+    #[ORM\Column]
+    private array $post_meta_title = [];
 
-    #[ORM\Column(type: 'string', length: 255, nullable: true)]
-    private $post_meta_title;
+    #[ORM\Column(nullable: true)]
+    private ?array $post_meta_desc = null;
 
-    #[ORM\Column(type: 'string', length: 255, nullable: true)]
-    private $post_meta_desc;
-
-    #[ORM\Column(length: 255, nullable: true)]
-    private ?string $post_meta_title_en = null;
-
-    #[ORM\Column(length: 255, nullable: true)]
-    private ?string $post_meta_desc_en = null;
-
-    #[ORM\Column(length: 255, nullable: true)]
-    private ?string $post_name_en = null;
+    #[ORM\Column(length: 255)]
+    private ?string $post_url = null;
 
     #[ORM\Column(length: 255, nullable: true)]
     private ?string $post_thumb = null;
 
-    #[ORM\Column]
-    private ?bool $status = null;
+    #[ORM\Column(nullable: true)]
+    private ?bool $online = null;
 
     #[ORM\Column]
     private ?\DateTimeImmutable $created_at = null;
@@ -49,17 +42,29 @@ class PostsList
     #[ORM\Column]
     private ?\DateTimeImmutable $updated_at = null;
 
+    #[ORM\ManyToOne(inversedBy: 'postsLists')]
+    #[ORM\JoinColumn(nullable: false)]
+    private ?User $author = null;
+
+    #[ORM\OneToMany(mappedBy: 'post', targetEntity: MenuLink::class)]
+    private Collection $menuLinks;
+
+    public function __construct()
+    {
+        $this->menuLinks = new ArrayCollection();
+    }
+
     public function getId(): ?int
     {
         return $this->id;
     }
 
-    public function getPostName(): ?string
+    public function getPostName(): array
     {
         return $this->post_name;
     }
 
-    public function setPostName(string $post_name): self
+    public function setPostName(array $post_name): static
     {
         $this->post_name = $post_name;
 
@@ -71,81 +76,9 @@ class PostsList
         return $this->post_url;
     }
 
-    public function setPostUrl(string $post_url): self
+    public function setPostUrl(string $post_url): static
     {
         $this->post_url = $post_url;
-
-        return $this;
-    }
-
-    public function getPostId(): ?string
-    {
-        return $this->post_id;
-    }
-
-    public function setPostId(string $post_id): self
-    {
-        $this->post_id = $post_id;
-
-        return $this;
-    }
-
-    public function getPostMetaTitle(): ?string
-    {
-        return $this->post_meta_title;
-    }
-
-    public function setPostMetaTitle(string $post_meta_title): self
-    {
-        $this->post_meta_title = $post_meta_title;
-
-        return $this;
-    }
-
-    public function getPostMetaDesc(): ?string
-    {
-        return $this->post_meta_desc;
-    }
-
-    public function setPostMetaDesc(?string $post_meta_desc): self
-    {
-        $this->post_meta_desc = $post_meta_desc;
-
-        return $this;
-    }
-
-    public function getPostMetaTitleEn(): ?string
-    {
-        return $this->post_meta_title_en;
-    }
-
-    public function setPostMetaTitleEn(?string $post_meta_title_en): self
-    {
-        $this->post_meta_title_en = $post_meta_title_en;
-
-        return $this;
-    }
-
-    public function getPostMetaDescEn(): ?string
-    {
-        return $this->post_meta_desc_en;
-    }
-
-    public function setPostMetaDescEn(?string $post_meta_desc_en): self
-    {
-        $this->post_meta_desc_en = $post_meta_desc_en;
-
-        return $this;
-    }
-
-    public function getPostNameEn(): ?string
-    {
-        return $this->post_name_en;
-    }
-
-    public function setPostNameEn(?string $post_name_en): self
-    {
-        $this->post_name_en = $post_name_en;
 
         return $this;
     }
@@ -155,21 +88,57 @@ class PostsList
         return $this->post_thumb;
     }
 
-    public function setPostThumb(?string $post_thumb): self
+    public function setPostThumb(?string $post_thumb): static
     {
         $this->post_thumb = $post_thumb;
 
         return $this;
     }
 
-    public function isStatus(): ?bool
+    public function getPostContent(): array
     {
-        return $this->status;
+        return $this->post_content;
     }
 
-    public function setStatus(bool $status): self
+    public function setPostContent(array $post_content): static
     {
-        $this->status = $status;
+        $this->post_content = $post_content;
+
+        return $this;
+    }
+
+    public function getPostMetaTitle(): array
+    {
+        return $this->post_meta_title;
+    }
+
+    public function setPostMetaTitle(array $post_meta_title): static
+    {
+        $this->post_meta_title = $post_meta_title;
+
+        return $this;
+    }
+
+    public function getPostMetaDesc(): ?array
+    {
+        return $this->post_meta_desc;
+    }
+
+    public function setPostMetaDesc(?array $post_meta_desc): static
+    {
+        $this->post_meta_desc = $post_meta_desc;
+
+        return $this;
+    }
+
+    public function getAuthor(): ?User
+    {
+        return $this->author;
+    }
+
+    public function setAuthor(?User $author): static
+    {
+        $this->author = $author;
 
         return $this;
     }
@@ -179,7 +148,7 @@ class PostsList
         return $this->created_at;
     }
 
-    public function setCreatedAt(\DateTimeImmutable $created_at): self
+    public function setCreatedAt(\DateTimeImmutable $created_at): static
     {
         $this->created_at = $created_at;
 
@@ -191,9 +160,51 @@ class PostsList
         return $this->updated_at;
     }
 
-    public function setUpdatedAt(\DateTimeImmutable $updated_at): self
+    public function setUpdatedAt(\DateTimeImmutable $updated_at): static
     {
         $this->updated_at = $updated_at;
+
+        return $this;
+    }
+
+    public function isOnline(): ?bool
+    {
+        return $this->online;
+    }
+
+    public function setOnline(?bool $online): static
+    {
+        $this->online = $online;
+
+        return $this;
+    }
+
+    /**
+     * @return Collection<int, MenuLink>
+     */
+    public function getMenuLinks(): Collection
+    {
+        return $this->menuLinks;
+    }
+
+    public function addMenuLink(MenuLink $menuLink): static
+    {
+        if (!$this->menuLinks->contains($menuLink)) {
+            $this->menuLinks->add($menuLink);
+            $menuLink->setPost($this);
+        }
+
+        return $this;
+    }
+
+    public function removeMenuLink(MenuLink $menuLink): static
+    {
+        if ($this->menuLinks->removeElement($menuLink)) {
+            // set the owning side to null (unless already changed)
+            if ($menuLink->getPost() === $this) {
+                $menuLink->setPost(null);
+            }
+        }
 
         return $this;
     }
