@@ -17,7 +17,8 @@ import * as Vue from 'vue';
 
 import { ScrollWeb } from './smoothScroll';
 import { Parallax } from './parallax';
-import { ImgConverter } from './imgConverter';
+import { createApp } from 'vue';
+import AOS from 'aos';
 
 // Variables
 // -----------------------------------------------
@@ -31,8 +32,11 @@ const values = {
 
 // Instantieur
 // -----------------------------------------------
-document.addEventListener('DOMContentLoaded', () => {
-    Vue.createApp({}).mount('#website');
+document.addEventListener('DOMContentLoaded', function(){
+    createApp({
+    //     Configuration de votre application Vue.js
+    }).mount('#website');
+    AOS.init();
     scrollWeb();
     parallax();
 });
