@@ -17,9 +17,32 @@ class PagesAdminFormType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('page_name', TextType::class, [
-                'label' => 'Nom de la page'
+            
+            // Contenu en FR
+            ->add('page_name_fr', TextType::class, [
+                'label' => 'Nom de la page',
+                'mapped' => false,
             ])
+            ->add('page_content_fr', CKEditorType::class, [
+                'label' => 'Contenu de la page (FR)',
+                'mapped' => false,
+                'attr' => [
+                    // Ajoutez une classe CSS pour identifier le champ de texte
+                    'class' => 'grapesjs-editor',
+                ],
+            ])
+            ->add('page_meta_title_fr', TextType::class, [
+                'label' => 'Balise Meta Title (FR)',
+                'required' => false,
+                'mapped' => false
+            ])
+            ->add('page_meta_desc_fr', TextareaType::class, [
+                'label' => 'Balise Meta Description (FR)',
+                'required' => false,
+                'mapped' => false
+            ])
+
+            // Global
             ->add('page_url', TextType::class, [
                 'label' => 'URL de la page',
                 'required' => false,
@@ -35,30 +58,7 @@ class PagesAdminFormType extends AbstractType
                 'label' => 'Page fixe',
                 'required' => false, 
             ])
-            ->add('page_content', CKEditorType::class, [
-                'label' => 'Contenu de la page (FR)',
-                'mapped' => false
-            ])
-            ->add('page_meta_title', TextType::class, [
-                'label' => 'Balise Meta Title (FR)',
-                'required' => false,
-            ])
-            ->add('page_meta_desc', TextareaType::class, [
-                'label' => 'Balise Meta Description (FR)',
-                'required' => false,
-            ])
-            ->add('page_content_en', CKEditorType::class, [
-                'label' => 'Contenu de la page (EN)',
-                'mapped' => false
-            ])
-            ->add('page_meta_title_en', TextType::class, [
-                'label' => 'Balise Meta Title (EN)',
-                'required' => false,
-            ])
-            ->add('page_meta_desc_en', TextareaType::class, [
-                'label' => 'Balise Meta Description (EN)',
-                'required' => false,
-            ])
+
             ->add('page_submit', SubmitType::class, [
                 'label' => 'Enregistrer'
             ])
