@@ -19,43 +19,27 @@ class PostsAdminFormType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('post_name', TextType::class, [
-                'label' => "Nom de l'article (FR)"
+            // FR
+            ->add('post_name_fr', TextType::class, [
+                'label' => "Nom de l'article (FR)",
+                'mapped' => false
             ])
-            ->add('status', ChoiceType::class, [
-                'label' => "Visibilité de l'article",
-                'choices' => [
-                    "En brouillon" => false,
-                    "Publié" => true
-                ]
-            ])
-            ->add('post_content', CKEditorType::class, [
+            ->add('post_content_fr', CKEditorType::class, [
                 'label' => "Contenu de l'article (FR)",
                 'mapped' => false
             ])
-            ->add('post_meta_title', TextType::class, [
+            ->add('post_meta_title_fr', TextType::class, [
                 'label' => "Meta Title de l'article (FR)",
-                'required' => false
-            ])
-            ->add('post_meta_desc', TextareaType::class, [
-                'label' => "Meta Description de l'article (FR)",
-                'required' => false
-            ])
-            ->add('post_name_en', TextType::class, [
-                'label' => "Nom de l'article (EN)"
-            ])
-            ->add('post_content_en', CKEditorType::class, [
-                'label' => "Contenu de l'article (EN)",
+                'required' => false,
                 'mapped' => false
             ])
-            ->add('post_meta_title_en', TextType::class, [
-                'label' => "Meta Title de l'article (EN)",
-                'required' => false
+            ->add('post_meta_desc_fr', TextareaType::class, [
+                'label' => "Meta Description de l'article (FR)",
+                'required' => false,
+                'mapped' => false
             ])
-            ->add('post_meta_desc_en', TextareaType::class, [
-                'label' => "Meta Description de l'article (EN)",
-                'required' => false
-            ])
+
+            // GLOBAL
             ->add('post_thumb', DropzoneType::class, [
                 'label' => "Image de l'article",
                 'required' => false,
@@ -64,6 +48,13 @@ class PostsAdminFormType extends AbstractType
             ->add('post_url', HiddenType::class, [
                 'label' => "URL de l'article (Optionnel)",
                 'required' => false
+            ])
+            ->add('online', ChoiceType::class, [
+                'label' => "Visibilité de l'article",
+                'choices' => [
+                    "En brouillon" => false,
+                    "Publié" => true
+                ]
             ])
             ->add('submit', SubmitType::class, [
                 'label' => "Enregistrer"
