@@ -11,7 +11,7 @@ use Symfony\Component\HttpFoundation\Request;
 
 class PagesService extends AbstractController{
     
-    function PageManager(ManagerRegistry $doctrine, Request $request, bool $newPage, String $page_id = null)
+    function PageManager(ManagerRegistry $doctrine, Request $request, bool $newPage, string $page_id = null)
     {
         $entityManager = $doctrine->getManager();
 
@@ -21,17 +21,15 @@ class PagesService extends AbstractController{
             $page = new PagesList();
         } else {
             $page = $entityManager->getRepository(PagesList::class)->findOneBy(['page_id' => $page_id]);
-            if(!$page) {
+            if (!$page) {
                 throw $this->createNotFoundException("Aucune page n'a été trouvée");
             }
         }
-
 
         // INITIALISATION DU FORMULAIRE
         // --------------------------------------------------------
         $form = $this->createForm(PagesAdminFormType::class, $page);
         $form->handleRequest($request);
-
 
         // ENVOI DU FORMULAIRE
         // --------------------------------------------------------
@@ -39,9 +37,13 @@ class PagesService extends AbstractController{
             // Récupération des données du formulaire
             $page = $form->getData();
 
+            // Création du nom
+            $name = [$form->get('page_name_fr')->getData()];
+            $page->setPageName($name);
+
             // Création du slug
             $slugify = new Slugify();
-            $slugName = $slugify->slugify($form->get('page_name')->getData());
+            $slugName = $slugify->slugify($name[0]);
             $slugUrl = $slugify->slugify($form->get('page_url')->getData());
 
             // Création de l'ID Page
@@ -49,29 +51,20 @@ class PagesService extends AbstractController{
                 $page->setPageId($slugName);
             }
 
-            // Création de l'URL
-            if ($form->get('page_url')->getData()) {
-                $page->setPageUrl($slugUrl);
-            } else {
-                if ($newPage) {
-                    $page->setPageUrl($slugName);
-                } else {
-                    $page->setPageUrl($page->setPageUrl($page->getPageUrl()));
-                }
-            }
+            // Création / Modification de l'URL
+            $page->setPageUrl(empty($form->get('page_url')->getData()) ? ($newPage ? $slugName : $page->getPageUrl()) : $slugUrl);
 
-            // Création du Meta Title
-            $metaTitle = $form->get('page_meta_title')->getData();
-            $page->setPageMetaTitle($metaTitle ? $metaTitle : $form->get('page_name')->getData());
+            // Création / Modification du Meta Title
+            $metaTitle = [$form->get('page_meta_title_fr')->getData() ?: $name[0]];
+            $page->setPageMetaTitle([$metaTitle]);
 
-            // Création / Modification du fichier TWIG
-            $pageFileName = $page->getPageId() . '.html.twig';
-            $file = fopen("../templates/webpages/pages/fr/" . $pageFileName, 'w');
-            $file_en = fopen("../templates/webpages/pages/en/" . $pageFileName, 'w');
-            fwrite($file, $form->get('page_content')->getData());
-            fclose($file);
-            fwrite($file_en, $form->get('page_content_en')->getData());
-            fclose($file_en);
+            // Création / Modification du Meta Desc
+            $metaDesc = [$form->get('page_meta_desc_fr')->getData() ?: ''];
+            $page->setPageMetaDesc($metaDesc);
+
+            // Création / Modification du contenu
+            $pageContent = [htmlspecialchars($form->get('page_content_fr')->getData()) ?: "Contenu à ajouter"];
+            $page->setPageContent($pageContent);
 
             // Envoi des données vers la BDD
             $entityManager->persist($page);
